added test cases to existing spec
